requirements.txt
flask==2.0.1
# SQLAlchemy==1.4.25
python-decouple==3.4
mysqlclient==2.0.3
Flask-SQLAlchemy==2.5.1